+++ src/lib/layout/search/SearchResultItemBody.tsx
@@ -1,62 +1,53 @@
 import { Flex, Text } from "@chakra-ui/react";
 
 import { PrimaryNameMark } from "lib/components/PrimaryNameMark";
-import type {
-  ResultMetadata,
-  SearchResultType,
-} from "lib/services/searchService";
+import type { SearchResult } from "lib/services/searchService";
 
 interface SearchResultItemBodyProps {
-  metadata: ResultMetadata;
-  value: string;
-  type: SearchResultType;
+  result: SearchResult;
   isAccountAddress: boolean;
 }
 
 export const SearchResultItemBody = ({
-  metadata,
-  value,
-  type,
+  result,
   isAccountAddress,
 }: SearchResultItemBodyProps) => {
-  const normalizedIcnsValue = value.endsWith(`.${metadata.icns.bech32Prefix}`)
-    ? value
-    : `${value}.${metadata.icns.bech32Prefix}`;
-
-  if (isAccountAddress && metadata.initiaUsername.address)
+  if (isAccountAddress && !!result.metadata?.initiaUsername)
     return (
       <Flex direction="column">
-        <Text variant="body2">{metadata.initiaUsername.username}</Text>
+        <Text variant="body2">{result.metadata.initiaUsername}</Text>
         <Flex
           gap={{ base: 0, md: 1 }}
           direction={{ base: "column", md: "row" }}
         >
           <Text variant="body2" color="text.dark" wordBreak="break-all">
-            {metadata.initiaUsername.username
-              ? `(${metadata.initiaUsername.address})`
-              : metadata.initiaUsername.address}
+            {result.metadata.initiaUsername
+              ? `(${result.value})`
+              : result.value}
           </Text>
           <Text
             variant="body2"
             fontWeight={{ base: "auto", md: 500 }}
             color="text.disabled"
           >
-            – {type}
+            – {result.type}
           </Text>
         </Flex>
       </Flex>
     );
 
-  if (isAccountAddress && metadata.icns.icnsNames?.primaryName)
+  if (isAccountAddress && !!result.metadata?.icns)
     return (
       <Flex direction="column">
         <Flex gap={1}>
           <Flex gap={1} align="center">
             <PrimaryNameMark />
-            <Text variant="body2">{metadata.icns.icnsNames.primaryName}</Text>
+            <Text variant="body2">
+              {result.metadata.icns.icnsNames.primaryName}
+            </Text>
           </Flex>
-          {value !== metadata.icns.address &&
-            normalizedIcnsValue !== metadata.icns.icnsNames?.primaryName && (
+          {result.value !== result.metadata.icns.address &&
+            result.value !== result.metadata.icns.icnsNames.primaryName && (
               <Text
                 variant="body2"
                 _before={{
@@ -66,7 +57,7 @@ export const SearchResultItemBody = ({
                   mr: 1,
                 }}
               >
-                {normalizedIcnsValue}
+                {result.value}
               </Text>
             )}
         </Flex>
@@ -75,16 +66,16 @@ export const SearchResultItemBody = ({
           direction={{ base: "column", md: "row" }}
         >
           <Text variant="body2" color="text.dark" wordBreak="break-all">
-            {metadata.icns.icnsNames.primaryName
-              ? `(${metadata.icns.address})`
-              : metadata.icns.address}
+            {result.metadata.icns.icnsNames.primaryName
+              ? `(${result.metadata.icns.address})`
+              : result.metadata.icns.address}
           </Text>
           <Text
             variant="body2"
             fontWeight={{ base: "auto", md: 500 }}
             color="text.disabled"
           >
-            – {type}
+            – {result.type}
           </Text>
         </Flex>
       </Flex>
@@ -92,22 +83,23 @@ export const SearchResultItemBody = ({
 
   if (
     isAccountAddress ||
-    type === "NFT Collection Address" ||
-    type === "Validator Address" ||
-    type === "Transaction Hash" ||
-    type === "Module Path"
+    result.type === "NFT Address" ||
+    result.type === "NFT Collection Address" ||
+    result.type === "Validator Address" ||
+    result.type === "Transaction Hash" ||
+    result.type === "Module Path"
   )
     return (
       <Flex gap={{ base: 0, md: 1 }} direction="column">
         <Text variant="body2" wordBreak="break-all">
-          {value}
+          {result.value}
         </Text>
         <Text
           variant="body2"
           fontWeight={{ base: "auto", md: 500 }}
           color="text.disabled"
         >
-          – {type}
+          – {result.type}
         </Text>
       </Flex>
     );
@@ -115,14 +107,14 @@ export const SearchResultItemBody = ({
   return (
     <Flex gap={1}>
       <Text variant="body2" wordBreak="break-all">
-        {value}
+        {result.value}
       </Text>
       <Text
         variant="body2"
         fontWeight={{ base: "auto", md: 500 }}
         color="text.disabled"
       >
-        – {type}
+        – {result.type}
       </Text>
     </Flex>
   );