+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1054](https://github.com/alleslabs/celatone-frontend/pull/1054) Improve main search bar to show both address and address as username
+
 ### Bug fixes
 
 - [#1050](https://github.com/alleslabs/celatone-frontend/pull/1050) Fix gap between network subsection